+++ app/libraries/Theme/Setups/Customizer/AwesomePosts/Controls/Number.php
@@ -14,7 +14,7 @@ final class Number extends AbstractControl
         $this->id = $awesome_posts->settings['Number']->id;
 
         $this->args['type'] = 'select';
-        $this->args['label'] = \esc_html__('Number or posts', 'my-theme');
+        $this->args['label'] = \esc_html__('Number or posts', 'jentil-theme');
         $this->args['choices'] = \array_combine(($num = \range(0, 6, 1)), $num);
     }
 }