+++ app/libraries/Theme/Setups/Scripts/CustomizePreview.php
@@ -21,7 +21,7 @@ final class CustomizePreview extends AbstractScript
     {
         parent::__construct($theme);
 
-        $this->id = "{$this->app->theme->stylesheet}-customize-preview";
+        $this->id = "{$this->app->meta['slug']}-customize-preview";
     }
 
     public function run()
@@ -66,15 +66,14 @@ final class CustomizePreview extends AbstractScript
      */
     public function enqueue()
     {
+        $file_system = $this->app->utilities->fileSystem;
         $file = '/dist/scripts/customize-preview.min.js';
 
         \wp_enqueue_script(
             $this->id,
-            $this->app->utilities->fileSystem->themeDir('url', $file),
+            $file_system->themeDir('url', $file),
             ['customize-preview'],
-            \filemtime(
-                $this->app->utilities->fileSystem->themeDir('path', $file)
-            ),
+            \filemtime($file_system->themeDir('path', $file)),
             true
         );
     }