+++ app/libraries/Theme/Setups/Styles/Gutenberg.php
@@ -12,7 +12,7 @@ final class Gutenberg extends AbstractStyle
     {
         parent::__construct($theme);
 
-        $this->id = "{$this->app->theme->stylesheet}-gutenberg";
+        $this->id = "{$this->app->meta['slug']}-gutenberg";
     }
 
     public function run()
@@ -25,19 +25,17 @@ final class Gutenberg extends AbstractStyle
      */
     public function enqueue()
     {
-        if (\is_rtl()) {
-            $file = '/dist/styles/gutenberg-rtl.min.css';
-        } else {
-            $file = '/dist/styles/gutenberg.min.css';
-        }
+        $file_system = $this->app->utilities->fileSystem;
+
+        $file = \is_rtl() ?
+            '/dist/styles/gutenberg-rtl.min.css' :
+            '/dist/styles/gutenberg.min.css';
 
         \wp_enqueue_style(
             $this->id,
-            $this->app->utilities->fileSystem->themeDir('url', $file),
+            $file_system->themeDir('url', $file),
             [$this->app->parent->setups['Styles\Gutenberg']->id],
-            \filemtime(
-                $this->app->utilities->fileSystem->themeDir('path', $file)
-            )
+            \filemtime($file_system->themeDir('path', $file))
         );
     }
 }