+++ assets/scripts/customize-preview.ts
@@ -1,40 +1,15 @@
-/// <reference path='./global.d.ts' />
+/// <reference path='./customize-preview/module.d.ts' />
 
-((_wp: WP, $: JQueryStatic): void => {
-    'use strict'
+import { Base } from './customize-preview/base'
 
-    const { customize } = _wp
+import { AwesomePostsHeading } from './customize-preview/awesome-posts-heading'
+import { BackgroundColor } from './customize-preview/background-color'
+import { BackgroundImage } from './customize-preview/background-image'
 
-    customize(myThemeAwesomePostsHeadingModId, (from: () => void): void => {
-        from.bind((to: string): void => {
-            $('#awesome-posts .heading').html(to)
-        })
-    })
+const previews = [
+    new AwesomePostsHeading(jQuery, wp, myThemeAwesomePostsHeadingModId),
+    // new BackgroundColor(jQuery, wp),
+    // new BackgroundImage(jQuery, wp),
+]
 
-    /**
-     * You may need these, if your theme adds 'custom-background'
-     * support.
-     *
-     * Corresponds to PHP method: `Setups\Background::addBodyClasses()`
-     */
-
-    // customize('background_color', (from: () => void): void => {
-    //     from.bind((to: string): void => {
-    //         if (-1 === ['#fff', '#ffffff'].indexOf(to)) {
-    //             $('body').removeClass('no-background-color').addClass('has-background-color')
-    //         } else {
-    //             $('body').removeClass('has-background-color').addClass('no-background-color')
-    //         }
-    //     })
-    // })
-
-    // customize('background_image', (from: () => void): void => {
-    //     from.bind((to: string): void => {
-    //         if (to) {
-    //             $('body').removeClass('no-background-image').addClass('has-background-image')
-    //         } else {
-    //             $('body').removeClass('has-background-image').addClass('no-background-image')
-    //         }
-    //     })
-    // })
-})(wp, jQuery)
+jQuery.each(previews, (_, preview: Base): void => preview.run())