+++ tests/unit/libraries/Theme/Setups/Translations/CoreTest.php
@@ -17,12 +17,7 @@ class CoreTest extends AbstractTestCase
         $add_action = FunctionMocker::replace('add_action');
 
         $translation = new Core(Stub::makeEmpty(AbstractChildTheme::class, [
-            'theme' => new class {
-                function get()
-                {
-                    return 'my-theme';
-                }
-            }
+            'meta' => ['text_domain' => 'theme'],
         ]));
 
         $translation->run();
@@ -41,15 +36,9 @@ class CoreTest extends AbstractTestCase
 
         $theme = Stub::makeEmpty(AbstractChildTheme::class, [
             'utilities' => Stub::makeEmpty(Utilities::class),
-            'theme' => new class {
-                function get()
-                {
-                    return 'my-theme';
-                }
-            }
+            'meta' => ['text_domain' => 'theme', 'domain_path' => '/path'],
         ]);
 
-        $theme->utilities = Stub::makeEmpty(Utilities::class);
         $theme->utilities->fileSystem = Stub::makeEmpty(FileSystem::class, [
             'themeDir' => function (string $type, string $append) {
                 return "/var/www/themes/my-theme{$append}";
@@ -63,7 +52,7 @@ class CoreTest extends AbstractTestCase
         $load->wasCalledOnce();
         $load->wasCalledWithOnce([
             $translation->textDomain,
-            '/var/www/themes/my-theme/lang'
+            '/var/www/themes/my-theme/path'
         ]);
     }
 }