+++ CHANGELOG.md
@@ -1,8 +1,29 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+## [0.5.0] - 2020-02-08
+
+### Added
+- Add `.security.txt`
+- Add suport for PHP 7.3 and 7.4
+- Set up JS module bundling with [rollup js](https://rollupjs.org)
+- Set up [broser-sync](https://www.browsersync.io)
+- Add `.gitattributes`
+- Add `Procfile`
+- Add `Dockerfile`
+
+### Changed
+- Rename theme back to `jentil-theme`
+- Rename `LICENSE.md` to `LICENSE`
+- Upgrade gulp to version 4
+- Update scripts to use JS modules
+
+### Removed
+- Remove HTML microdata
 
 ## 0.4.0 - 2018-09-27